Make '--outputFile' automatically set '--full'.
This is to prevent irrecoverable data loss in the case of human error. I.e.: "sqs-grep -q Queue -a -o file.txt --delete".
It is VERY easy to forget to set the --full flag and then later try to use the dump file and find out it is not in  the right format.